pyproject.toml
[tool.poetry]
name = "home-work-poetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.14.1"
isort = "^5.13.2"
black = "^24.10.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = '.venv'

[tool.isort]
line_length = 119

[tool.black]
line-length = 119
exclude = '''(\.git)'''
